requirements.txt
Django==5.1.2
playwright==1.48.0
ruff==0.7.3
loguru==0.7.2
altair==5.4.1
pandas==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
requests==2.32.3